Add fish controls, tweak game engine
It is now possible to control the fish with the arrows. It uses a
"realistic" physics engine using EOM in viscous fluid.
For game, changed the way we draw things (now passes the screen to
the "Objet" class and it handles the drawing). Event handling
changed to control the fish. A bit messy as is, maybe there is a
easier way to implement the controls.