Refine animation with rotation. Change some graphics.
Need to fix the collision on the borders, there is a problem because
we determine the fish angle with the speed. However the speed is set
to 0 on the borders. If we don't set the speed to 0, then the problem
is that we get slightly "stuck" on the borders becausewe need to first
overcome the negative speed before getting "unstuck".